refactor: RequestBody 분리
